feat: create builder class instead of index.js in utils
