+++ schematics/cloud/api/appsync/data/expectedNewFiles.json
@@ -1,4 +1,1 @@
-[
-  "/cloud/lib/api/appSyncCdkConstruct.ts",
-  "/cloud/lib/api/appSyncCdkConstruct.test.ts"
-]
+["/cloud/lib/api/appSyncCdkConstruct.ts", "/cloud/lib/api/appSyncCdkConstruct.test.ts"]
